% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Classify observations}
\usage{
classify(x, theta)
}
\arguments{
\item{x}{A \code{matrix} of A) observations where rows corresponds to
features and columns to studies or B) class probabilities with}

\item{theta}{A list of parameters for the full model as described in
\code{\link{rtheta}}. If \code{theta} is supplied, \code{x} are assumed to
be observations (A). If \code{theta} is missing, \code{x} are assumed to be
probabilites (B).}
}
\value{
A integer vector of class numbers with length equal to the number of
  rows in \code{x}.
}
\description{
Classify observations according to the maximum a posterior probabilites.
}
\examples{
set.seed(4821)
theta <- rtheta(d = 2, m = 3)
x <- matrix(runif(75), 25, 3)
classify(x, theta)
}
\seealso{
\code{\link{get.prob}}
}
