% Generated by roxygen2 (4.0.1): do not edit by hand
\name{colSds}
\alias{colSds}
\title{Compute row and column standard deviations}
\usage{
colSds(x)
}
\arguments{
\item{x}{A numeric matrix of size \code{n} times \code{m}}
}
\value{
\code{rowSds} returns a numeric vector of length \code{n}.
}
\description{
The \code{rowSds} and \code{colSds} respectively computes the
standard deviations of the rows and columns of the given matrix.
}
\examples{
x <- matrix(rnorm(50), 10, 5)
rowSds(x)
}
\author{
Anders Ellern Bilgrau <abilgra''umath.aau.dk>
}
\seealso{
\code{\link{rowMeans}}, \code{\link{colMeans}}
}

