% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cummean.R
\name{cummean}
\alias{cummean}
\title{Cumulative mean values}
\usage{
cummean(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
A vector of length \code{length(x)} with the cumulative mean. The
\code{i}'th entry \code{cummean(x)[i]} equals \code{mean(x[1:i])}.
}
\description{
Returns a vector whose \code{i}'th element is the cumulative mean
(arithmetic mean) of the \code{i}'th first elements of the argument.
}
\examples{
x <- sort(rnorm(100))
GMCM:::cummean(x)
}
\seealso{
\code{\link{cumsum}}
}
\author{
Anders Ellern Bilgrau <anders.ellern.bilgrau@gmail.com>
}
\keyword{internal}
