% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgmcm.loglik.R, R/dgmm.loglik.R,
%   R/dgmm.loglik.marginal.R, R/pgmm.marginal.R, R/qgmm.marginal.R
\name{dgmcm.loglik}
\alias{dgmcm.loglik}
\alias{dgmm.loglik}
\alias{dgmm.loglik.marginal}
\alias{pgmm.marginal}
\alias{qgmm.marginal}
\title{Probability, density, and likelihood functions of the Gaussian mixture
(copula) model}
\usage{
dgmcm.loglik(theta, u, marginal.loglik = FALSE, ...)

dgmm.loglik(theta, z, marginal.loglik = FALSE)

dgmm.loglik.marginal(theta, x, marginal.loglik = TRUE)

pgmm.marginal(z, theta)

qgmm.marginal(u, theta, res = 1000, spread = 5, rule = 2)
}
\arguments{
\item{theta}{A list parameters as described in \code{\link{rtheta}}.}

\item{u}{A matrix of (estimates of) realizations from the GMCM where each
row corresponds to an observation.}

\item{marginal.loglik}{Logical. If \code{TRUE}, the marginal log-likelihood
functions for each multivariate observation (i.e. the log densities) are
returned. In other words, if \code{TRUE} the sum of the marginal
likelihoods is not computed.}

\item{\dots}{Arguments passed to \code{qgmm.marginal}.}

\item{z}{A matrix of realizations from the latent process where each row
corresponds to an observation.}

\item{x}{A matrix where each row corresponds to an observation.}

\item{res}{The resolution at which the inversion of \code{qgmm.marginal} is
done. Default is 1000.}

\item{spread}{The number of marginal standard deviations from the marginal
means the \code{pgmm.marginal} is to be evaluated on.}

\item{rule}{The extrapolation rule used in \code{\link{approxfun}}.}
}
\value{
The returned value depends on the value of \code{marginal.loglik}.
  If \code{TRUE}, the non-summed marginal likelihood values are returned. If
  \code{FALSE}, the scalar sum log-likelihood is returned.

  \code{dgmcm.loglik}: As above, with the GMCM density.

  \code{dgmm.loglik}: As above, with the GMM density.

  \code{dgmm.loglik.marginal}: As above, where the j'th element is evaluated
  in the j'th marginal GMM density.

  \code{pgmm.marginal}: A matrix where the (i,j)'th entry is the (i,j)'th
  entry of \code{z} evaluated in the jth marginal GMM density.

  \code{qgmm.marginal}: A matrix where the (i,j)'th entry is the (i,j)'th
  entry of \code{u} evaluated in the inverse jth marginal GMM density.
}
\description{
Marginal and simultaneous cumulative distribution, log probability density,
and log-likelihood functions of the Gaussian mixture model (GMM) and
Gaussian mixture copula model (GMCM) and the relevant inverse marginal
quantile functions.
}
\details{
\code{qgmm.marginal} distributes approximately \code{res} points around the
cluster centers according to the mixture proportions in \code{theta$pie} and
evaluates \code{pgmm.marginal} on these points. An approximate inverse of
\code{pgmm.marginal} function is constructed by linear interpolation of the
flipped evaluated coordinates.
}
\examples{
set.seed(1)
data <- SimulateGMCMData(n = 10)
u <- data$u
z <- data$z
print(theta <- data$theta)

GMCM:::dgmcm.loglik(theta, u, marginal.loglik = FALSE)
GMCM:::dgmcm.loglik(theta, u, marginal.loglik = TRUE)

GMCM:::dgmm.loglik(theta, z, marginal.loglik = FALSE)
GMCM:::dgmm.loglik(theta, z, marginal.loglik = TRUE)

GMCM:::dgmm.loglik.marginal(theta, z, marginal.loglik = FALSE)
GMCM:::dgmm.loglik.marginal(theta, z, marginal.loglik = TRUE)

GMCM:::pgmm.marginal(z, theta)
GMCM:::qgmm.marginal(u, theta)
}
\author{
Anders Ellern Bilgrau <anders.ellern.bilgrau@gmail.com>
}
\keyword{internal}
