% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dmvnormal}
\alias{dmvnormal}
\alias{rmvnormal}
\title{Multivariate Gaussian probability density function}
\usage{
dmvnormal(x, mu, sigma)
}
\arguments{
\item{x}{A p times k matrix of quantiles. Each rows correspond to a
realization from the density and each column corresponds to a dimension.}

\item{n}{The number of observations to be simulated.}

\item{mu}{The mean vector of dimension k.}

\item{sigma}{The variance-covariance matrix of dimension k times k.}
}
\value{
\code{dmvnorm} returns a 1 by p matrix of the probability densities
corresponding to each row of x.

\code{rmvnorm} returns a p by k matrix of observations from at multivariate
normal distribution with the given mean \code{mu} and covariance
\code{sigma}. Each row corresponds to an observation.
}
\description{
Fast simulation and evalutation of multivariate Gaussian probability
densities.
}
\details{
\code{dmvnormal} functions similarly to \code{dmvnorm} from the
\code{mvtnorm}-package.
}
\examples{
GMCM:::dmvnormal(x = matrix(rnorm(300), 100, 3),
                 mu = 1:3,
                 sigma = diag(3))
}
\author{
Anders Ellern Bilgrau
}
\seealso{
\code{dmvnorm} and \code{rmvnorm} in the \code{mvtnorm}-package.
}
\keyword{~kwd1}
\keyword{~kwd2}

