% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.theta.R
\name{plot.theta}
\alias{plot.theta}
\title{Plotting method for "theta" objects}
\usage{
\method{plot}{theta}(theta, which.dims = c(1L, 2L), n.sd = 2,
  add.means = TRUE, ..., add.ellipses = FALSE)
}
\arguments{
\item{which.dims}{An integer vector of length 2 choosing which two dimensions
to plot.}

\item{n.sd}{An integer choosing the number of standard deviations in each
dimension to determine the plotting window.}

\item{...}{Arguments passed to \code{contour}.}

\item{add.ellipses}{logical. If TRUE, ellipses outlining 95\% confidence
regions for each component are added in the bivariate multivariate
distribution defined by theta and \code{which.dims}.}

\item{x}{An object of class \code{theta}.}
}
\value{
Plots via the \code{contour} function. Invisibly returns a list with
  x, y, z coordinates that is passed to contour.
}
\description{
Visualizes the chosen dimensions of the theta object graphically by the GMM
density and possibly the individual gaussian components.
}
\examples{
set.seed(5)
theta <- rtheta(d = 3, m = 4)
plot(theta)
plot(theta, col = "blue", asp = 1, add.means = FALSE)
plot(theta, col = "blue", asp = 1, add.means = TRUE)
plot(theta, which.dims = c(3L, 2L), asp = 1)
plot(theta, asp = 1, n.sd = 3, add.ellipses = TRUE,
     nlevels = 40, axes = FALSE)
}
\author{
Anders Ellern Bilgrau <anders.ellern.bilgrau@gmail.com>
}
