% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.theta.R
\name{print.theta}
\alias{print.theta}
\title{Print method for theta class}
\usage{
\method{print}{theta}(x, ...)
}
\arguments{
\item{x}{A \code{theta} object. See \code{\link{rtheta}}.}

\item{...}{Arguments to be passed to subsequent methods.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for theta class
}
\examples{
theta <- rtheta()
print(theta)
}
\author{
Anders Ellern Bilgrau <anders.ellern.bilgrau@gmail.com>
}
