% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGMCM.R
\name{runGMCM}
\alias{runGMCM}
\title{Run the GMCM shiny application}
\usage{
runGMCM(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[shiny]{runApp}}.}
}
\value{
Retuns nothing (usually). See \code{\link[shiny]{runApp}}.
  Exit or stop the app by interrupting R.
}
\description{
Function for starting a local instance of the GMCM shiny application.
The online application is found at \url{https://gmcm.shinyapps.io/GMCM/}.
}
\examples{
\dontrun{
runGMCM()
runGMCM(launch.browser = FALSE, port = 1111)
# Open browser and enter URL http://127.0.0.1:1111/
}
}
\seealso{
\code{\link[shiny]{runApp}}
}
