% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Uhat}
\alias{Uhat}
\title{Fast ranking function}
\usage{
Uhat(x)
}
\arguments{
\item{x}{A matrix of observations to be ranked. Rows correspond to features
and columns to experiments.}
}
\value{
A matrix with the same dimensions as \code{x} of the scaled ranks.
}
\description{
Function for computing the scaled ranks for each column of the input matrix.
I.e. the values are ranked column-wise and divided by \code{nrow(x) + 1}.
}
\examples{
data <- SimulateGMMData()
par(mfrow = c(1,2))
plot(data$z, xlab = expression(z[1]), ylab = expression(z[2]))
plot(Uhat(data$z), xlab = expression(hat(u)[1]), ylab = expression(hat(u)[2]))
}
\author{
Anders Ellern Bilgrau (abilgrau@math.aau.dk)
}
\seealso{
\code{\link{SimulateGMMData}}, \code{\link{SimulateGMCMData}}
}

