% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.theta.R
\name{as.theta}
\alias{as.theta}
\title{Coerce a list to a theta object}
\usage{
as.theta(x)
}
\arguments{
\item{x}{A theta-like object that can be coerced.}
}
\value{
A theta object. See \code{\link{rtheta}}.
}
\description{
At the moment, only the class is added.
}
\examples{
m <- 2
d <- 2
x <- list(m = m,
          d = d,
          pie = c(0.5, 0.5),
          mu = list(comp1=rep(0,d), comp2=rep(1,d)),
          sigma = list(comp1=diag(d), comp2=diag(d)))
print(x)
theta <- as.theta(x)
print(theta)
}
