% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Classify observations}
\usage{
classify(x, theta)
}
\arguments{
\item{x}{A \code{matrix} of observations where rows corresponds to features
and columns to studies.}

\item{theta}{A list of parameters for the full model as described in
\code{\link{rtheta}}.}
}
\value{
A integer vector of class numbers.
}
\description{
Classify observations according to the maximum a posterior probabilites.
}
\examples{
set.seed(4821)
theta <- rtheta(d = 2, m = 3)
x <- matrix(runif(75), 25, 3)
classify(x, theta)
}
\seealso{
\code{\link{get.prob}}
}
