% Generated by roxygen2 (4.0.2): do not edit by hand
\name{colSds}
\alias{colSds}
\alias{rowSds}
\title{Compute row and column standard deviations}
\usage{
colSds(x)

rowSds(x)
}
\arguments{
\item{x}{A numeric matrix of size \code{n} times \code{m}}
}
\value{
\code{colSds} returns a numeric vector of length \code{m}.

\code{rowSds} returns a numeric vector of length \code{n}.
}
\description{
The \code{rowSds} and \code{colSds} respectively computes the
  standard deviations of the rows and columns of the given matrix.
}
\examples{
x <- matrix(rnorm(50), 10, 5)
GMCM:::colSds(x)
y <- matrix(rnorm(50), 10, 5)
GMCM:::rowSds(y)
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}
\seealso{
\code{\link{rowMeans}}, \code{\link{colMeans}}
}
\keyword{internal}

