% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rtheta}
\alias{rtheta}
\title{Get random parameters for the Gaussian mixture (copula) model}
\usage{
rtheta(m = 3, d = 2)
}
\arguments{
\item{m}{The number of components in the mixture.}

\item{d}{The dimension of the mixture distribution.}
}
\value{
A named list of parameters with the 4 elements:
  \item{m}{The number of components in the mixture. Default is 3.}
  \item{d}{The dimension of the mixture distribution. Default is 2.}
  \item{pie}{A numeric vector of mixture proportions which sums to one.}
  \item{mu}{A list of mean vectors for each component.}
  \item{comp}{A list of variance-covariance matrices for each component.}
}
\description{
The function generates a random set parameters for the Gaussian mixture
model (GMM) and Gaussian mixture copula model (GMCM).
}
\note{
The function \code{\link{is.theta}} checks whether or not \code{theta}
  is in the correct format.
}
\examples{
GMCM:::rtheta()
GMCM:::rtheta(d = 5, m = 2)
}
\author{
Anders Ellern Bilgrau (abilgrau@math.aau.dk)
}
\seealso{
\code{\link{is.theta}}
}

