% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGMCM.R
\name{runGMCM}
\alias{runGMCM}
\title{Run the GMCM shiny app}
\usage{
runGMCM(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[shiny]{runApp}}.}
}
\value{
Retuns nothing (usually). See \code{\link[shiny]{runApp}}.
  Exit or stop the app by interrupting R.
}
\description{
Function for locally running the GMCM shiny application.
}
\examples{
\dontrun{
runGMCM()
runGMCM(launch.browser = FALSE, port = 1111)
}
}
\seealso{
\code{\link[shiny]{runApp}}
}
