% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.theta.R
\name{summary.theta}
\alias{summary.theta}
\title{Summary method for theta class}
\usage{
\method{summary}{theta}(object, ...)
}
\arguments{
\item{object}{A \code{theta} object. See \code{\link{rtheta}}.}

\item{...}{arguments to be passed to subsequent methods}
}
\value{
Invisibly returns \code{x}.
}
\description{
Summary method for theta class
}
\examples{
theta <- rtheta()
summary(theta)
}
\author{
Anders Ellern Bilgrau <anders.ellern.bilgrau@gmail.com>
}
